refactor: 큐카드 데이터 get hook 리팩토링
